refactor: Remove unused 'drag' parameter from popup menus (rv-ui)
This commit removes the 'drag' parameter from several popup menu
composables, as it is no longer used within their logic. This
simplifies the composable signatures and reduces unnecessary
parameter passing.

The following composables have had the 'drag' parameter removed:
- PrivateApplicationInfoPopup
- PrivateApplicationInfoMenu
- PrivateShortcutInfoMenu
- PrivateShortcutInfoMenuItem

Additionally, the LaunchedEffect related to drag state in
PrivateShortcutInfoMenuItem has been removed as it is no longer
relevant.

This change is a cleanup effort to remove dead code and improve
maintainability.
